pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "actualvoids-discord-bot"
version = "0.1.0"
description = "Discord bot and API bridge for processing crypto withdrawal requests from a Minecraft plugin."
authors = [
  {name = "actualvoids"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.3.2",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=1.10.13,<2.0.0",
  "python-dotenv>=1.0.0",
  "aiohttp>=3.9.0",
  "yarl>=1.9.0"
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
